+++ CHANGELOG.md
@@ -182,6 +182,7 @@ parts of your citations, using the following CSS classes:
   from both link and link reference), the tooltips (which accurately show link
   previews), rendering (which accounts for link labels), to any action (such as
   copying or removing a link) (#5142).
+- **Feature**: Zettlr now supports loading BibLaTeX libraries as well (#460).
 - **Change**: Snippets: The `$FILENAME` variable now does not contain the file
   extension anymore. Users who also want the extension should update their
   snippets to `$FILENAME$EXTENSION` (#4191).